third_party/WebKit/LayoutTests/platform/win/web-animations-api/animation-playbackRate-expected.txt

This is a testharness.js-based test.
PASS animation.currentTime should be constant if there is a 0 playback rate 
Harness: the test ran to completion.
